[CP] Fix cmp fail on segment number>256 (#1783)
* Fix cmp fail on segment number>256
* Fix cur_segment_length not calculate startcode on Windows

+++ _studio/mfx_lib/pxp/include/mfx_pxp_video_accelerator.h
@@ -38,7 +38,7 @@ namespace UMC
         mfxHDL GetPXPCtxHdl() { return m_PXPCtxHdl; }
         void SetPXPParams(mfxHDL pxpParams) { m_PXPParams = pxpParams; }
         mfxHDL GetPXPParams() { return m_PXPParams; }
-        uint8_t m_curSegment;
+        uint32_t m_curSegment;
     private:
         mfxHDL m_PXPCtxHdl;
         mfxHDL m_PXPParams;
